+++ client/package.json
@@ -11,9 +11,7 @@
     "format": "prettier --write \"src/**/*.{ts,tsx,css}\""
   },
   "dependencies": {
-    "@apollo/client": "^3.3.13",
     "cookies-js": "^1.2.3",
-    "graphql": "^15.5.0",
     "lint-staged": "^10.5.4",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
